+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 * `SitemapBuilder::getMethods()` returns a collection and it's now public;
 * `StaticPage::all()` returns a collection;
 * `AppView::userbar()` always returns the userbar as array;
+* `Plugin::all()` method now takes the `mecms_core` option, which excludes
+    plugins automatically requested by MeCms;
 * fixed `TestSuite` classes, `MenuBuilderHelper` and some template files, improved
     to use with other plugins;
 * fixed old bug on loading theme plugins;